🔨 Dependency Upgrades 1. Upgrade to Gradle 9.1.0 2. Upgrade to Spring Boot 3.5.6 2. Upgrade to Portainer 2.34.0
⭐ New Features
1. 实体类改为 kt，并给字段添加上 @Column 与 @Comment 注解，代码优化
2. table.sql 优化，主键 bigint 改为够用的 INTEGER

+++ src/main/java/com/simonvonxcvii/turing/service/impl/NimbusJwtServiceImpl.kt
@@ -96,7 +96,7 @@ class NimbusJwtServiceImpl(
                     .audience(listOf(username))
                     // 设置过期时间 （exp） 声明，该声明标识不得接受 JWT 进行处理的时间或之后的时间。
                     // 形参: expiresAt – 不得接受 JWT 进行处理的时间或之后的时间
-                    .expiresAt(now.plusSeconds(securityProperties.expires))
+                    .expiresAt(now.plusSeconds(securityProperties.expires.toLong()))
                     // 设置不早于 （nbf） 声明，该声明标识不得接受 JWT 进行处理的时间。
                     // 形参: notBefore——不得接受 JWT 进行处理的时间
                     .notBefore(now)
